feat(ios): refactor backup settings UI by replacing ExportImportSheet with dedicated ExportSheet and ImportSheet components
